api/requirements.txt

flask
flask-jwt-extended
flask-limiter
python-dotenv
argon2-cffi
gunicorn
psycopg2-binary
